[FEATURE] Support to inject Resources into beans
Due to the dynamic representation of resources in this framework, repositories and session beans are obsolete now.

However. In some circumstances the use of these resources in other beans (such as session beans or cdi beans) is necessary.
To not call this resources internally via http, we provide a solution within this feature to inject resources into beans and use them as needed:

Suppose there is an example resource:
~~~~~
@Resource( /* ... */ )
@GET( /* ... */ )
public interface MyTestResource {
  String getGreetings();
}
~~~~~

Now it is possible to inject an instance of the `Resources` interface in another bean:
~~~~~
@Stateless
public class MyTestBean {
  @Inject
  Resources<MyTestResource> testResources;

  //...
}
~~~~~

The generic parameter of this `Resources` interface is an interface, tagged with `@Resource` annotation, that is used inside the MyTestBean.

With this injected instance you can programmatically access instances of `MyTestResource` by selecting the resource and invoking the corresponding methods for the HTTP-Calls as defined in the ReST Resource:

~~~~~
    MyTestResource instance = testResources.select("greetingsFor", "Robert").GET();
    System.out.println(instance.getGreetings());
~~~~~

With this possibility, the `Resources` API just follows CDI and reuse the pattern stated with CDI `@Inject Instance<MyBeanType> beanInstances`.

See the javadoc and the source for more details.